requirements.txt
ShopifyAPI>=12.3.0
PyYAML>=6.0.1